test(smt): use confined process fixture
