+++ tests/boundary/process/test_cvc5_worker_command_profile.py
@@ -116,3 +116,35 @@ def test_cvc5_worker_still_rejects_commands_outside_the_supported_profile(
         "protocol": "jacobian.cvc5-worker/v1",
         "error_code": "CVC5_COMMAND_OUTSIDE_PROFILE",
     }
+
+
+def test_smt_parser_state_guards_survive_optimized_python() -> None:
+    script = """
+from jacobian.contracts import smt as contract_smt
+from jacobian_checkers import smt as checker_smt
+
+if __debug__:
+    raise SystemExit("optimized Python was not enabled")
+
+for module, opener_name in (
+    (contract_smt, "_open_smtlib_command"),
+    (checker_smt, "_smtlib_open_paren"),
+):
+    setattr(module, opener_name, lambda depth, direct_atoms: (1, None))
+    try:
+        module._top_level_commands("(x")
+    except ValueError as exc:
+        if str(exc) != "SMT-LIB parser state is inconsistent":
+            raise
+    else:
+        raise SystemExit(f"{module.__name__} accepted inconsistent parser state")
+"""
+
+    completed = subprocess.run(
+        [sys.executable, "-O", "-c", script],
+        check=False,
+        capture_output=True,
+        text=True,
+    )
+
+    assert completed.returncode == 0, completed.stderr
